Реализовал воспроизведение звука при отправке/приеме сообщений
