+++ .idea/encodings.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Encoding" defaultCharsetForPropertiesFiles="UTF-8" addBOMForNewFiles="with BOM">
+  <component name="Encoding" defaultCharsetForPropertiesFiles="UTF-8">
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/java/Controller.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/Controller/Controller.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/Controller/ControllerCheckIn.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/Controller/ControllerLogin.java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java/Main.java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java/MainApp.java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/java/Utils.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/Utils/Utils.java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources/sound" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources/sound/out.wav" charset="UTF-8" />
