+++ pom.xml
@@ -12,7 +12,6 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     </properties>
-
     <build>
 
 <!--        <resources>-->
@@ -64,10 +63,8 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.5</version>
             </plugin>
-
         </plugins>
     </build>
-
 
     <dependencies>
         <dependency>
@@ -97,6 +94,12 @@
         <dependency>
             <groupId>org.openjfx</groupId>
             <artifactId>javafx-web</artifactId>
+            <version>14.0.1</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-media</artifactId>
             <version>14.0.1</version>
         </dependency>
