+++ src/main/java/Controller/ControllerCheckIn.java
@@ -1,0 +1,5 @@
+package Controller;
+
+public class ControllerCheckIn {
+
+}
