+++ src/main/java/Controller/ControllerLogin.java
@@ -1,0 +1,34 @@
+package Controller;
+
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.PasswordField;
+import javafx.scene.control.TextField;
+
+public class ControllerLogin {
+
+    @FXML
+    private Button button_check_in;
+
+    @FXML
+    private TextField login;
+
+    @FXML
+    private PasswordField password;
+
+    @FXML
+    private Button button_entry;
+
+    @FXML
+    void putBottonEntry(ActionEvent event) {
+
+    }
+
+    @FXML
+    void putButtonCheckIn(ActionEvent event) {
+
+    }
+
+}
+
