--- src/main/java/Utils.java
@@ -1,27 +1,0 @@
-import javax.sound.sampled.AudioInputStream;
-import javax.sound.sampled.AudioSystem;
-import javax.sound.sampled.Clip;
-import java.io.File;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-public class Utils {
-    static String dateTimeCreate() {
-        Date data = new Date();
-        DateFormat df = new SimpleDateFormat("HH:mm:ss");
-        String datetime = df.format(data);
-        return datetime;
-    }
-
-    public static void playSound(String sound) {
-        try {
-            AudioInputStream audioInputStream = AudioSystem.getAudioInputStream(new File(sound));
-            Clip clip = AudioSystem.getClip();
-            clip.open(audioInputStream);
-            clip.start();
-        } catch (Exception var4) {
-            var4.printStackTrace();
-        }
-    }
-}
