+++ src/main/resources/FXML/Login.fxml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controller.ControllerLogin">
+   <children>
+      <AnchorPane layoutX="26.0" layoutY="-16.0" prefHeight="200.0" prefWidth="200.0">
+         <children>
+            <AnchorPane layoutX="-26.0" layoutY="14.0" prefHeight="119.0" prefWidth="600.0" style="-fx-background-color: #778899; -fx-border-radius: 10;">
+               <children>
+                  <Button fx:id="button_check_in" layoutX="467.0" layoutY="20.0" mnemonicParsing="false" onAction="#putButtonCheckIn" prefHeight="25.0" prefWidth="119.0" text="Регистрация" />
+               </children></AnchorPane>
+            <Label layoutX="154.0" layoutY="24.0" text="BolTalCa" textFill="#fffdfd99">
+               <font>
+                  <Font name="Comic Sans MS" size="56.0" />
+               </font>
+            </Label>
+            <Label layoutX="191.0" layoutY="141.0" text="Авторизация" textFill="#14131399">
+               <font>
+                  <Font name="Comic Sans MS" size="27.0" />
+               </font>
+            </Label>
+         </children>
+      </AnchorPane>
+      <TextField fx:id="login" layoutX="108.0" layoutY="208.0" prefHeight="25.0" prefWidth="385.0" promptText="Логин" style="-fx-background-color: white; -fx-border-color: silver;">
+         <font>
+            <Font size="15.0" />
+         </font>
+      </TextField>
+      <PasswordField fx:id="password" layoutX="108.0" layoutY="276.0" prefHeight="25.0" prefWidth="385.0" promptText="Пароль" style="-fx-background-color: white; -fx-border-color: silver;">
+         <font>
+            <Font size="15.0" />
+         </font>
+      </PasswordField>
+      <Button fx:id="button_entry" layoutX="274.0" layoutY="342.0" mnemonicParsing="false" onAction="#putBottonEntry" text="Button" />
+   </children>
+</AnchorPane>
